requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.7.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.6
flatbuffers==24.12.23
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
jax==0.4.30
jaxlib==0.4.30
joblib==1.4.2
keras==2.12.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
Pillow==9.5.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-macos==2.12.0
termcolor==2.5.0
tflearn==0.5.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.21.0